Separate component with redux container
